// src/app/model/Product.ts
export class Product {
    
  conventionprice: Number;
  deliveryfees: Number;
  depth: Number;
  description: string;
  freedelivery: Number;
  height: Number
  isactif: Number;
  isdeleted: Number;
  isnew: Number;
  ispromotion: Number;
  length: Number;
  name: string;
  payments: string;
  priceht: Number;
  pricetype: string;
  
  publicprice: Number;
  reference: string;
  rejectioncause: string
  serialized: Number
  sku: string
  sn: string
  soldstock: Number
  specs: string
  stock: Number;
  terms: Number;
  validation: Number;
  waranty: Number;
  weight: Number;
  withdelivery: Number
  brandid :Number;
  idcategorie:Number;
  idvat:Number;
}